feat: Finish authorization code handling
Due to unforeseen ownership issues, the API was reworked so that the
`CliOAuth` does not take ownership of the `oauth2::Client` struct, or
perform the token exchange. This library is then relegated to an
intermediate role in the middle of the PKCE flow.

The library has been fleshed out to the point of saving the
authorization code and state, printing out the authorization URL, and
being able to compare the state from the auth code request to the
expected state (what was passed to the authorization URL).